* REFACTOR controls test
The aim is to compartmentalize aspects of the player control and actions into separate player state objects that utilize a common interface. The player update function would run the update method of whatever PlayerState interface is assigned to the property on the main player class. As the player changes button presses the states cause the player to switch which PlayerState is assigned.
Also attempting to split out values related to character settings into a separate object, for organization purposes. Ideally would like settings to be exposed through Unity editor so they can be manipulated during runtime, but intial pass at this fails because I'm referencing this in the player class via an instance of this class.